ML/receipt-scanner/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
httpx==0.28.1
python-dotenv==1.1.0
pydantic==2.11.1
paddlepaddle==3.0.0
paddleocr==2.9.1
Pillow>=10.0.0
numpy>=1.24.0
# uvicorn app.main:app --reload --port 8000